backend/requirements.txt
Django~=5.0.7
openai~=1.36.1
python-dotenv~=1.0.1
requests~=2.32.3
beautifulsoup4~=4.12.3
django-cors-headers~=4.4.0
djangorestframework~=3.15.2
python-decouple~=3.8
django-celery-beat~=2.6.0